added logout and drop down menu
