+++ frontend/src/components/Navigation/Navigation.css
@@ -1,0 +1,3 @@
+.hidden{
+    display: none
+}
